pyproject.toml
[project]
name = "rai"
version = "1.0.0"
description = "RAI is a next-gen CLI framework to automate the creation of intelligent agents and teams for cybersecurity and offensive security operations"
license = {text = "MIT"}
authors = [
    { name = "D. Sanjai Kumar", email = "bughunterz0047@gmail.com" } 
]
maintainers = [
    { name = "D. Sanjai Kumar", email = "bughunterz0047@gmail.com" }, { name = "Gokul.V", email = "th3gokul@gmail.com"}
]
keywords = ["rai", "ai-agents", "multi-agent", "framework", "ethical hacking", "penetration testing", "AI shell", "ctf", "llm", "agno", "LLM Team", "AI Security"]
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries",
    "Topic :: Utilities",
]

dependencies = [
    "agno>=1.4.3",
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.18",
    "anthropic>=0.50.0",
    "appdirs>=1.4.4",
    "art>=6.5",
    "asyncio>=3.4.3",
    "azure-ai-inference>=1.0.0b9",
    "boto3>=1.38.6",
    "cohere>=5.15.0",
    "colorama>=0.4.6",
    "google-genai>=1.13.0",
    "groq>=0.23.1",
    "httpx>=0.28.1",
    "ibm-watsonx-ai>=1.3.13",
    "litellm>=1.67.5",
    "mcp>=1.6.0",
    "mistralai>=1.7.0",
    "nest-asyncio>=1.6.0",
    "ollama>=0.4.8",
    "openai>=1.76.2",
    "prompt-toolkit>=3.0.51",
    "pyyaml>=6.0.2",
]

[project.urls]
Homepage = "https://github.com/RevoltSecurities/RAI" 
Documentation = "https://github.com/RevoltSecurities/RAI/wiki"
Repository = "https://github.com/RevoltSecurities/RAI"
Issues = "https://github.com/RevoltSecurities/RAI/issues"


[project.scripts]
rai = "rai.rai:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
